+++ src/main/java/me/sisyphusj/community/app/post/domain/PostDetailVO.java
@@ -8,10 +8,10 @@ import lombok.Getter;
 public class PostDetailVO {
 
 	// 작성자 ID
-	private int userId;
+	private long userId;
 
 	// 게시글 고유 ID
-	private int postId;
+	private long postId;
 
 	// 작성자
 	private String author;
